requirements.txt
flask
pynamodb
zappa
python-jose
